% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MYplot.R
\name{MYplot}
\alias{MYplot}
\title{Annotates Dusa's XYplot of fsQCA with Inclusion and Coverage measures
using various methods (C, Ragin, T. Haesebrouck and F. Veri)}
\usage{
MYplot(
  expr,
  out,
  dat,
  jitter = TRUE,
  enhance = FALSE,
  Veri = TRUE,
  model = FALSE
)
}
\arguments{
\item{expr}{Expression in SOP or DNF form, in quotes}

\item{out}{Name of the outcome variable, in quotes}

\item{dat}{The calibrated data}

\item{jitter}{Passed to XYplot function}

\item{enhance}{Passed to XYplot function}

\item{Veri}{TRUE to outline Ambiguous and Non-Ambiguous cases areas on plot}

\item{model}{Passed to XYplot function}
}
\description{
Annotates Dusa's XYplot of fsQCA with Inclusion and Coverage measures
using various methods (C, Ragin, T. Haesebrouck and F. Veri)
}
