% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/VeriCov.R
\name{VeriCov}
\alias{VeriCov}
\title{Computes sufficiency coverage and consistency parameters 
by Ragin (R), Haesebrouck (H) and Veri (V) methods}
\usage{
VeriCov(
  model,
  out,
  dat,
  jitter = TRUE,
  enhance = FALSE,
  sol = 1,
  tablerows = "both",
  tablecols = "Veri",
  plot = FALSE
)
}
\arguments{
\item{model}{Either QCA minimized model object or one or more SOP expressions in quotes}

\item{out}{Name of the outcome variable, in quotes}

\item{dat}{The calibrated data}

\item{jitter}{Passed to XYplot function}

\item{enhance}{Passed to XYplot function}

\item{sol}{is the solution number (for model objects)}

\item{tablerows}{is 'terms', 'solution' or 'both'}

\item{tablecols}{is 'none', 'Ragin', 'Veri', 'both' or 'details'}

\item{plot}{is FALSE, TRUE or 'Veri'}
}
\description{
Computes sufficiency coverage and consistency parameters 
by Ragin (R), Haesebrouck (H) and Veri (V) methods
}
